chances are, this will fix the x86 build

+++ src/cascadia/TerminalControl/TermControl.cpp
@@ -2351,7 +2351,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
     // - Gets the internal taskbar state value
     // Return Value:
     // - The taskbar state of this control
-    const size_t TermControl::TaskbarState() const noexcept
+    const uint64_t TermControl::TaskbarState() const noexcept
     {
         return _core.TaskbarState();
     }
@@ -2360,7 +2360,7 @@ namespace winrt::Microsoft::Terminal::Control::implementation
     // - Gets the internal taskbar progress value
     // Return Value:
     // - The taskbar progress of this control
-    const size_t TermControl::TaskbarProgress() const noexcept
+    const uint64_t TermControl::TaskbarProgress() const noexcept
     {
         return _core.TaskbarProgress();
     }
